Moved context into a separate file and turned carpdp into a server component
