+++ src/app/(main)/[productType]/components/EditVehiclePopover.tsx
@@ -7,7 +7,7 @@ import {
   PopoverContent,
   PopoverTrigger,
 } from '@/components/ui/popover';
-import { CarSelectionContext } from './CarPDP';
+import { CarSelectionContext } from '@/contexts/CarSelectionContext';
 import { useContext, useState } from 'react';
 import { IProductData } from '../../utils';
 import { useParams } from 'next/navigation';
